Cast to unsigned type after checking that max is positive. This gets rid of a compiler warning without compromising correctness

+++ json.c
@@ -311,7 +311,7 @@ static int buffer_grow(json_parser *parser)
 	if (max > 0 && parser->buffer_size == max)
 		return JSON_ERROR_DATA_LIMIT;
 	newsize = parser->buffer_size * 2;
-	if (max > 0 && newsize > max)
+	if (max > 0 && newsize > (uint32_t)(max))
 		newsize = max;
 
 	ptr = parser_realloc(parser, parser->buffer, newsize * sizeof(char));
